refactor(cmd): split endpoint_diagnostics.go into cohesive files
endpoint_diagnostics.go had grown to ~1370 lines spanning doctor, inventory,
collector-repair, and bundle concerns. Extract two cohesive, self-contained
clusters into their own files in package cmd:

- endpoint_repair.go: the collector service repair + rollback flow, the
  repairServiceManager/repairFileSnapshot/repairRollback types, and the
  package-level test-injection seams.
- endpoint_inventory.go: the inventory run/heartbeat/snapshot commands and their
  filter helpers, plus the inventoryResult/inventoryHeartbeatWriteResult types.

Move-only: no logic changes. endpoint_diagnostics.go drops to ~800 lines and now
holds doctor/checks/bundle/hooks. Because everything stays in package cmd,
visibility and the diagnostics test injection seams are unchanged.
